backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
apscheduler==3.10.4
spacy==3.7.4
dateparser==1.2.0
fastapi-mail==1.4.1
python-multipart==0.0.9
pydantic[email]==2.6.4
httpx==0.27.0